Refactor into common.ts to get rid of index import dep
